+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 ### Changed
 - Ampache API:
   * Action `get_indexes` supports also `type=song_artist`
+  * Actions `playlists` and `playlist` support argument `include`
   * Added fields `art` and `has_art` to the `podcast_episode` and `live_stream` result types
   * Added field `username` to the responses of `handshake` and `ping`
   * Added fields `md5`, `has_access`, `has_collaborate`, and `last_update` to the `playlist` results